tests/wpt/web-platform-tests/web-animations/interfaces/Animation/pending.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation.pending</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animation-pending">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

promise_test(t => {
  const div = createDiv(t);
  const animation = div.animate({}, 100 * MS_PER_SEC);

  assert_true(animation.pending);
  return animation.ready.then(() => {
    assert_false(animation.pending);
  });
}, 'reports true -> false when initially played');

promise_test(t => {
  const div = createDiv(t);
  const animation = div.animate({}, 100 * MS_PER_SEC);
  animation.pause();

  assert_true(animation.pending);
  return animation.ready.then(() => {
    assert_false(animation.pending);
  });
}, 'reports true -> false when paused');

</script>
</body>
